+++ invisibleroads_macros/exceptions.py
@@ -1,0 +1,10 @@
+class InvisibleRoadsError(Exception):
+    pass
+
+
+class BadURL(InvisibleRoadsError):
+    pass
+
+
+class BadRepository(InvisibleRoadsError):
+    pass